refactor fairchem calculator and pretrained
