+++ src/fairchem/core/__init__.py
@@ -10,7 +10,8 @@ from __future__ import annotations
 from importlib.metadata import PackageNotFoundError, version
 
 from fairchem.core._config import clear_cache
-from fairchem.core.common.calculator import FAIRChemCalculator
+from fairchem.core.calculate import pretrained_mlip
+from fairchem.core.calculate.ase import FAIRChemCalculator
 
 try:
     __version__ = version("fairchem.core")
@@ -18,4 +19,4 @@ except PackageNotFoundError:
     # package is not installed
     __version__ = ""
 
-__all__ = ["FAIRChemCalculator", "clear_cache"]
+__all__ = ["FAIRChemCalculator", "pretrained_mlip", "clear_cache"]
